HW4/google/requirements.txt
Flask==3.0.3
waitress==3.0.0
python-dotenv==1.0.1
authlib==1.3.1